fix(types): apeGmsh.opensees mypy-clean (178 -> 0), MYPY_BASELINE 0 + docs sweep
Pays the full mypy baseline down to zero and lowers the static-gates
ratchet to a hard gate. Two of the 178 were real API-surface bugs:

* ops.pattern.Plain/UniformExcitation(series=...) falsely rejected
  newer TimeSeries: the annotation was a concrete union frozen at
  Linear/Constant/Path/Trig/Pulse, so the Ricker wavelet + cyclic
  protocols (PR #558) were rejected for type-checked callers and the
  stage-pattern helpers (typed against the base) never type-checked.
  series= now accepts TimeSeries | str; runtime resolution unchanged.
* Quoted annotations naming Plain / ConstraintRecord / Any with no
  import anywhere blinded mypy at ~20 signatures.

Everything else is annotation/narrowing only, zero behavior change:
* type-abstract disabled package-wide: the _resolve(ref, base=...)
  convention only isinstance-checks base (never instantiates); the
  separate `abstract` code still catches direct instantiation (~70
  false positives removed).
* Renamed reused loop variables that pinned types across loops.
* Fail-loud guards on Optionals whose invariants are now stated in
  the message (split='parts' requires a buffered Tcl/Py emitter;
  fem_eid_to_ops_tag populated before emit; from_mander fcc/fl XOR).
* Removed stale type: ignore comments; precise container/dtype
  annotations; no-reexport imports moved to defining modules.

Docs sweep alongside:
* README: de-versioned the stale "What's new in viewers (v1.5.0)"
  callout (current version is v2.0.0).
* guide_loads.md: fixed broken TOC anchor (double hyphen vs the
  generated slug).
* docs/api/loads.md: autorefs updated pattern -> case (ADR 0051
  vocabulary); section retitled Cases.
* CHANGELOG: plane-wave-ssi link made docs-root-relative so mkdocs
  resolves it (matches the existing internal_docs link convention).
* mkdocs.yml: documented that mkdocs 1.6.1 has no 'error' level for
  anchor validation (review follow-up considered, not applicable).

Verification: mypy 0 errors / ruff clean on src/apeGmsh/opensees;
full curated suite 9412 passed; docs build down to 7 structural
warnings (deliberate internal_docs -> src ADR links).

+++ src/apeGmsh/opensees/material/_mander.py
@@ -34,7 +34,7 @@ __all__ = [
 
 def confined_strength(fc: float, fl: float) -> float:
     """Mander confined peak strength from effective lateral pressure ``fl``."""
-    return fc * (2.254 * (1.0 + 7.94 * fl / fc) ** 0.5 - 2.0 * fl / fc - 1.254)
+    return float(fc * (2.254 * (1.0 + 7.94 * fl / fc) ** 0.5 - 2.0 * fl / fc - 1.254))
 
 
 def confined_peak_strain(fc: float, fcc: float, eps_co: float) -> float:
@@ -69,7 +69,7 @@ def compression_backbone(
 
     def mander(eps: float) -> float:
         x = eps / eps_cc
-        return fcc * x * r / (r - 1.0 + x ** r)
+        return float(fcc * x * r / (r - 1.0 + x ** r))
 
     fc0 = 0.5 * fcc
     ec0 = fc0 / E   # elastic anchor: secant slope is exactly E
